refactor: move paragraph element retrieval functionality to separate function/file
