+++ src/paragraphFormatting/index.ts
@@ -10,25 +10,10 @@ import { registeredFormatters } from "./registeredFormatters";
  * @param pluginSettings - The settings for the Gospel Study plugin, used to determine which formatters are enabled.
  * @returns An array of formatted paragraph strings.
  */
-export function getFormattedParagraphs(document: Document, paragraphIdItems: string[], pluginSettings: GospelStudyPluginSettings): string[] {
-	const paragraphElements: Element[] = [];
-
+export function getFormattedParagraphs(paragraphElements: Element[], pluginSettings: GospelStudyPluginSettings): string[] {
 	const enabledFormatters = registeredFormatters.filter((formatter: ParagraphFormatter) => formatter.isEnabled(pluginSettings));
-
-	const elementsWithIds = Array.from(document.querySelectorAll("[id]"));
-
-	paragraphIdItems.forEach((idItem) => {
-		if (idItem.includes("-")) { // It is a range of paragraphs
-			const rangeOfParagraphs = paragraphRangeToParagraphElements(idItem, elementsWithIds);
-			paragraphElements.push(...rangeOfParagraphs);
-
-		} else { // It is a single paragraph
-			const paragraphElement = paragraphIdToParagraphElement(idItem, elementsWithIds);
-			paragraphElements.push(paragraphElement);
-		}
-	});
-
 	const paragraphs: string[] = [];
+
 	paragraphElements.forEach((paragraphElement) => {
 		const formattedParagraph = formatParagraph(paragraphElement.innerHTML, enabledFormatters);
 		paragraphs.push(formattedParagraph);
@@ -37,23 +22,6 @@ export function getFormattedParagraphs(document: Document, paragraphIdItems: str
 	return paragraphs;
 }
 
-
-/**
- * Finds and returns the paragraph element with the specified ID from a list of elements.
- *
- * @param idItem - The ID of the paragraph element to find.
- * @param elementsWithIds - An array of elements that contain IDs.
- * @returns The paragraph element with the specified ID.
- * @throws Will throw an error if no element with the specified ID is found.
- */
-function paragraphIdToParagraphElement(idItem: string, elementsWithIds: Element[]): Element {
-	const paragraphElement = elementsWithIds.find((element) => element.id === idItem);
-
-	if (!paragraphElement) {
-		throw new Error(`Invalid paragraph ID: ${idItem}`);
-	}
-	return paragraphElement;
-}
 
 /**
  * Formats a paragraph by applying various transformations based on the plugin settings.
@@ -70,25 +38,3 @@ export function formatParagraph(paragraph: string, enabledFormatters: ParagraphF
 	return paragraph;
 }
 
-/**
- * Extracts a range of paragraph elements from a list of elements based on the provided range string.
- *
- * @param range - A string representing the range of paragraph elements, in the format "startId-endId".
- * @param elementsWithIds - An array of elements, each with an `id` property.
- * @returns An array of elements that fall within the specified range.
- * @throws Will throw an error if either the start or end ID is not found in the elements array.
- */
-function paragraphRangeToParagraphElements(range: string, elementsWithIds: Element[]): Element[] {
-	const [startId, endId] = range.split("-");
-
-	const startElementIndex = elementsWithIds.findIndex((element) => element.id === startId);
-	const endElementIndex = elementsWithIds.findIndex((element) => element.id === endId);
-
-	if (startElementIndex === -1 || endElementIndex === -1) {
-		throw new Error(`Invalid paragraph range: ${range}`);
-	}
-
-	const paragraphElements = elementsWithIds.slice(startElementIndex, endElementIndex + 1);
-
-	return paragraphElements;
-}
